docs: add instructions to install zbarimg
The tests were failing on a fresh clone because `zbarimg` was not installed and the code invoking it essentially ignores the error.

https://github.com/votingworks/module-scan/blob/7d3e233408221f9374d2238fe4576594b0e0b9ae/src/interpreter.ts#L98-L102

This binary is provided by the `zbar` package on macOS (via Homebrew) or `zbar-tools` on Ubuntu. This commit updates the README to include instructions for installing `zbarimg`, which fixes the tests.

Fixes https://github.com/votingworks/module-scan/issues/21