refactor: clean up types in `isJSON` utils
This gets rid of `any` and replaces it with [`unknown`][unknown] as appropriate. Using `unknown` is better because it is the _least_ capable type rather than `any` which is the _most_ capable type. This commit changes `isString` and `isObject` to be [type predicates][type predicates] which allows TS to infer types of arguments based on the return value, so `unknown` can be implicitly cast to `string` or `object` when control flow is based on the results of `isString` and `isObject`. Note that `isString` is not really necessary since `typeof value === 'string'` does the same thing, but I left it to minimize changes to this file.

`passObject` now has a default value of `false` just to simplify its type from `boolean | undefined` to just `boolean` while keeping the API the same.

Using `.map((s: any) => isJSON(s)).reduce((prev: any, curr: any) => !!curr)` is not good for a few reasons:
- all those `any` type annotations are unnecessary
- using `Array#reduce` where `prev` is ignored is a hint that there's a more appropriate method to use, in this case `Array#every`
- using `!!` on array entries is redundant when they are guaranteed to be `boolean` values because they come from `isJSON`

[unknown]: https://devblogs.microsoft.com/typescript/announcing-typescript-3-0-rc-2/#the-unknown-type
[type predicates]: http://www.typescriptlang.org/docs/handbook/advanced-types.html#using-type-predicates