refactor: use node-quirc instead of zbarimg
Why? I have not found a way to get zbarimg to give me back the raw data from the QR code. Instead, it gives me the bytes interpreted as UTF-8. This isn't good, as round-tripping through UTF-8 is not guaranteed to retain the original data.

quirc is a C library for reading QR codes that seems to work quite well (https://github.com/dlbeer/quirc) and can give us the raw binary data. The node wrapper for it, node-quirc, has JPEG support on my fork which this commit references. When my pull request lands (https://github.com/kAworu/node-quirc/pull/10), we can refer to the upstream version again.

quirc is roughly the same speed as zbarimg in my testing and so far is able to read all the QR codes I've thrown at it.