fix: ensure reading from card with `readFromCard.py` works
This script failed when the bytes in either short or long value couldn't
be interpreted as a utf-8 string. Now, we print a string for each if we
can, but fall back to printing the raw bytes if needed. Additionally, if
the bytes _are_ a valid utf-8 string but that string is not printable,
use the inspected version of the string rather than printing with blanks
for unprintable characters.

Closes #38