feat: add a built-in screen reader
We've relied on ChromeVox for our screen reading functionality so far. Getting it to load and use the installed voices on Linux has been inconsistent at best. The problem scope is much smaller since we don't have to implement an entire screen reader, just enough functionality to support what we use. This turned out to be surprisingly easy.

The tricky parts were in deciding which components are responsible for what and making sure things happen in the right order. There were already several parts responsible for various things related to this:
- `gamepad` takes the "keyboard" input from the gamepad and immediately triggers events based on them.
- `FocusManager` watches the current route and, if it changes, refocuses after a delay.
- `LinkButton` handles setting history for route changes, but did so after a delay. The stated reason for this in the comments was that the tap on e.g. "Next" would carry over to the next route, but that no longer seems to be true, at least on macOS with a pointer. I suspect when we changed the way taps worked that it made that delay unnecessary, but this needs to be verified on a touch screen on Linux.
- `AriaScreenReader` is now used by `App` to respond to events delegated by `App` to read text aloud using the built-in `SpeechSynthesis` APIs.