feat: enable speech by default from speech-loader entry point
When speech-loader.html is able to load voices properly, it will load index.html with a query string parameter telling it to enable the screen reader right away.

The API needed to change to support this. Instead of swapping out `AriaScreenReader` instances with different text-to-speech implementations (null or real), we add the ability to mute TTS instances. This actually maps pretty well to the current volume control in the speech synthesis API which we may want to use in the future anyway. So now we keep the screen reader instance and just enable/disable it.

In addition, a number of tests were not waiting properly for the async methods they were calling. The tests worked anyway, but this fixes it to make the more correct.