fix: maybe address freezing issue
This causes VxMark to quit after being idle on the "Insert Card" screen for 5 minutes. This MUST be paired with some sort of respawn functionality. It will only quit if the machine ID does _not_ end in "0", so with machine IDs using sequential integers this means ~90% of machines will have this restart behavior and  ~10% will not.