fix: prevent importers from clobbering each other
In production we don't have multiple importers running in parallel, but in tests we do. This ensures that each importer has its own directory for ballot images and that they do not interfere with each other. Without this, the end-to-end tests would sometimes time out in watch mode due to never seeing the `add` events after the files were removed by a still-running test in another file.