fix: allow scanning ballots if we have enough information
Previously we would disallow scanning any ballots if we didn't have all templates for all ballot styles and all precincts and all pages loaded. That's probably still desireable for a real election, but for testing it's often useful to just load a single ballot style/precinct combination. This change allows us to do so.