feat: add parsing for election definition objects
There are lots of places we might need to do this, and passing around unvalidated data in JSON makes me nervous. Doing this already caught an issue with the definition: `shortTitle` was not listed as optional but it was in fact missing in the sample election definition. BMD already assumed it was optional, so I made it optional in the TS types.

I explored a few approaches to this, including building JSON Schema definitions from the TS types. For this approach I ended up leaving the types separately defined from the schemas even though you can extract the type of a schema using `z.TypeOf<MySchema>`. I did this in part because VS Code & TS show nicer names for the types (i.e. `Election` instead of `{ title: string; … }`. TS will help keep things in sync since we use the hand-written type definition as the return type of `parseElection`, whose returned value comes from the `Election` schema.

This does simplify the locale types a bit too since I couldn't find a way to express the previous types in Zod easily. I think this will be fine.