test: add test to ensure digits can be in file names
Beau pointed out what appeared to be an issue with the file name generation code at https://github.com/votingworks/election-manager/pull/41/files#r450598526. This turned out to not be a problem as it only replaces digits with dashes in the human-readable language names. Still, I was here so I added a test and realized that I no longer needed an optional chain property access before calling `replace`.