fix: improve the quality of retained ballot images
The main problem this is addressing is that we were only storing the normalized ballot images, which caused debugging to be pretty annoying when an image was causing problems. The original output of the scanner was lost, and instead a JPEG with fairly heavy compression artifacts remained.

To address this problem, this commit does two things:
- switch from JPEG to PNG as the preferred scanned image format
- retain both the normalized ballot image and the original

Why PNG? In my testing I found that the file size is smaller (~90KB for PNG vs ~900KB for JPEG) without an apparent loss in scanning accuracy. Of note, the PNGs generated by the scanner are have 1-bit color depth, so every pixel is either black or white. This does mean we're trusting the binarization done by the scanning stack rather than our own. I have no reason to suspect this will be a problem, but it is a difference from JPEG.

As part of this change, the expected format is more abstracted and is currently only hardcoded in scanner.ts and importer.ts. The scanning format may be given as a parameter to `FujitsuScanner`, and the normalized image format can be changed by updating a call to sharp's `.png()` to `.jpeg()` instead.