feat: detect more adjudication conditions
In addition to marginal marks, this commit adds support for detecting overvote, undervote, and write-in conditions. It currently does not act on them except to log the adjudication reason and that it was ignored. These will not be ignored in the future, and will be configurable by the end user at some point.