refactor: watch scanimage output for files to import
This replaces chokidar file watching with the `--batch-print` option to the `scanimage` CLI. It prints each scanned image file's path to stdout followed by a newline. I made the minimal changes required to pass the information back to the caller, providing a callback to `scanInto`.

This ended up having some significant effects in terms of testing, which is where the majority of changes in this commit go.