feat: scan ballot cards one at a time
This required a lot of restructuring. The scanner now scans one card at a time, yielding two paths to scanned image files. It uses the `--batch-prompt` option to `scanimage` which waits for a newline to be sent to `stdin` before scanning the next page (or writing the buffered back side of an already-scanned to a file).

There was a problem using `--batch-prompt` and `--batch-print` in combination with `--swskip`, which is that skipped images do not count toward the requested ballot pages. Therefore, if one side of a ballot card was skipped then the next sheet would be pulled into the feeder and the front side of that sheet would be given as the back side of the first document. To work around this, we no longer use `--swskip`.  Instead, we detect blank pages ourselves using the same threshold of 0.5% foreground (dark) pixels.

`Scanner`'s API changed to have a `scanSheet` method that returns an async generator, i.e. a generator that yields promises. This async generator had to be constructed manually (as opposed to using `async function*`) because of the mismatch between the push-based nature of IO streams and the pull-based nature of generators. To reconcile the two, I created a `DeferredQueue` class, which is an async queue that matches up the consumer and producer sides using the deferred promise pattern.