fix: re-create database if schema has changed
Changes are detected by creating a digest of the SQL file used to initialize the database. If it matches the `.digest` file associated with a database, it'll open and use the database as normal. If not, either because the digest file does not exist or because the value is different, the existing database will be backed up and a new one created in its place.