fix: ensure switching between test/live mode works
The underlying interpreter is taught how to read the templates either when a new ballot package is uploaded or on startup when loading template information from the database. When switching test/live mode, the interpreter was recreated but _without teaching it about the templates_. Therefore, no matter what it was given, it would always throw "Cannot scan ballot because not all required templates have been added". That is, every `interpretBallot` call would fail.

To fix this issue, this commit restores the config from the database after switching test mode. Because this takes a non-trivial amount of time, we need to update bsd to tell the user what's happening.

Closes votingworks/vxmail#230