fix: always use module-scan for QR code reading
Without this, hmpb-interpreter was using its default QR code readers, node-quirc and jsqr. With this change, it always uses node-quirc and zbarimg (via qrdetect). The only reason we were allowing hmpb-interpreter to do its own QR code reading rather than providing it the metadata we already have is that it used the QR code position to orient the image correctly. With this change, we should be able to rotate the image when we detect it is flipped and pass the rotated image and parsed metadata to hmpb-interpreter so that it no longer needs to read QR codes at all. That's a good follow-up change to this one.