refactor(workspace): enable yarn workspaces
## Why?
This is the first step in moving toward a monorepo, and will make the remaining steps easier. Using yarn workspaces allows multiple packages to share dependencies since yarn hoists shared packages to the root `node_modules`. It also will link workspace packages together, which is the thing that enables monorepo development.

https://classic.yarnpkg.com/en/docs/workspaces

## Approach
To use yarn workspaces, you create a `package.json` at the root of the project with `private: true` and a `workspaces` property whose value is an array of glob paths. For now I just reference the two existing top-level folders, but later I'll move these into a more extensible structure.

## Issues
If you do this naively by just creating the top-level `package.json` and running `yarn`,  you may run into problems. Specifically, yarn re-evaluates the semver ranges in all the `package.json` dependencies to determine which version of packages to install. This is because yarn now considers only a top-level `yarn.lock`, ignoring ones that may exist alongside the project `package.json` files. Since the vast majority of dependencies were in the `client` folder, I just moved the `yarn.lock` from `client` to the top level before running `yarn` so that the existing versions from that project would be used. Without this, there were a variety of problems trying to run the application.

Despite moving `yarn.lock` to preserve as many resolved versions as possible, there  was still an error booting the application: `proxy is not a function`. This error came from this line: `app.use(proxy('/convert', { target: 'http://localhost:3003/' }))`. The problem was that `client/src/setupProxy.js` required `http-proxy-middleware` without depending on it directly. This worked because yarn hoists non-conflicting packages to enable sharing, so it ended up in the top-level `node_modules`, enabling us to require it. However, we were implicitly relying on another transitive dependency to install it, `webpack-dev-server`, which wanted `http-proxy-middleware@0.19.1`. When we created the workspace, yarn saw that `prodserver/package.json` declared a dependency on `http-proxy-middleware@^1.0.0`. Because this version is not compatible with the one `webpack-dev-server` wants, yarn gave `webpack-dev-server` its own copy in its `node_modules` and hoisted the one our workspace depended on explicitly to the root. Since the API changed from exporting a function as `module.exports` to exporting named exports as an object, trying to call an object in the `proxy` variable as a function yields the error: `proxy is not a function`. To fix this, I changed `prodserver` to depend on the same version we were implicitly using before: `0.19.1`. Note that the `1.0.0` version was never even being used since `prodserver` was calling into `client` to actually require it! Another way to fix it would have been to update `setupProxy` to use the API from the newer version of `http-proxy-middleware`.

## Alternatives
The issue above is a problem of a particular `node_modules` layout, specifically caused by hoisting dependencies. Yarn hoists them in order to reduce duplication, which is a good thing, but at the expense of strictness/correctness. There is another alternative package manager client called pnpm that supports workspaces in a similar way to yarn, but uses a `node_modules` layout that allows sharing and is strict: https://pnpm.js.org/en/workspaces. It may be worth evaluating moving to that at some point, but this commit is just making the smallest change possible.