fix: boost signal from actually-marked targets
This reduces the radius that the mark score minimization algorithm can go to find a lower score. We've been seeing it start dead center, but then moving away in order to shrink the score. This is not what we want, and this is not going to fix that, but it will reduce the effect. Experimentally, a radius of 2 makes all the zero scores stay zero, but boosts the non-zero scores. In the future we may want to also "maximize" the score as well, then use both scores to determine what to do.