fix: find corners more reliably by backtracking
When looking for corners we try to work our way in from the bounding box corners looking at increasingly large distances along both the x and y directions. Once a run of black pixels of sufficient length is found, we consider the first black pixel of that run to be a corner. In addition, we ensure that we do not go beyond a distance that would imply greater skew than expected.

With this algorithm, we still ran into non-corner points treated as a corner, leading the homography to introduce more skew than was in the original image. To fix this, the changes in this commit attempt to backtrack when finding a potential corner. For example, from the top left we might find a "corner" near the left edge of the bounding box, but farther down than the real corner. From there, we now move diagonally NE to see whether that is also a black pixel. If it is, scan straight up until we find a white pixel. Repeat this process until the NE diagonal check finds a white pixel. The last black pixel we found is the new corner.