fix: prevent accidental overvotes due to homography error
Sometimes a homography algorithm must choose which points to use and which to ignore. Sometimes, this means ignoring the top left corner of a contest box. Doing so means that, when there's a lot of skew, the bubbles at the top of that contest do not line up well with the template. This can cause spurious overvotes.

Since we primarily care about the bubbles/targets, and they are on the left side of the contest box, we tell jsfeat's homography algorithm to consider only the left sides of the contest boxes, hopefully ensuring none are thrown away.