fix: find corners better yet again
This approach ensures that we only ever look at points as potential corners if they are on the edge of the contest box shape. The previous approach looked at the bounding box, but skew could cause other contest boxes or other things to interfere with finding the corners.

Now, we track the edges of shapes and use that to scan in both x and y directions from all corners, looking for edge points that are closest to the bounding box corner. Once we've found all the points that have the same distance, we pick the point that is as close in both the x and y directions as any of those points got.