perf: optimize converting an image to grayscale
This commit adds an alternate, faster version of the grayscale converter. It benchmarks it against the existing version and jsfeat's version. The new version is the fastest (labeled "one pixel at a time" below) on my machine:

```
RGBA to grayscale
one byte at a time x 16.86 ops/sec ±1.22% (46 runs sampled)
one pixel at a time x 20.16 ops/sec ±0.79% (37 runs sampled)
jsfeat x 7.44 ops/sec ±3.78% (22 runs sampled)
Fastest is one pixel at a time
```

The old one in the benchmark above is about 59ms vs the new one at 50ms. This isn't a huge savings, but this commit is mostly about introducing a means of testing performance improvements. The benchmark is added in this commit, but I'll remove it in the next.

I believe the reason this is faster is because I'm doing ~4x fewer load and store operations on the pixel array since I'm getting and setting one RGBA pixel at a time (4 bytes) vs one byte at a time.