feat(module-scan): add `retry-scan` command
This adds a CLI command to retry scanning previously-scanned sheets. You can either scan a subset using a filter like `--unreadable` or by specifying sheets by id, or rescan all of them with `--all`. By default it'll use the current database for module-scan, but you can provide an alternate database such as from a backup file. It'll print out a summary of all the changes between the original scan and the rescan, including a diff of the interpretation if the type is unchanged (e.g. the `votes` changed in a HMPB). This is configurable with `--diff-when`.

My plan is to add an `--out-db` option that will write the new interpretations to a new database file for easier analysis of the results when working with larger datasets.