fix(hmpb-interpreter): reject non-corresponding contests
This compares the area of the contests in the scanned ballot vs. the template. If they are too different, we bail on interpretation. This can happen when there is a fold that "breaks" a box, making it appear smaller than it really is.

In the future we'll do more to fix issues caused by folding, but for now this will properly reject such pages that may otherwise be interpreted incorrectly.