fix(hmpb-interpreter): handle fold lines better
One problem we had in November was that folds would create phantom lines in the scans that sometimes connected to the edges of contest boxes. When this happens, the bounding box for the found shape may be larger than the allowed contest box width even though the contest box itself is within the expected size range.

To address this, we use the width and height as determined by looking at the distance between the corners. If that fits the expected dimensions then we consider it a contest box, even if some part of it sticks out.