refactor(module-scan): introduce "workspaces"
We had a bunch of cruft built up around the directories we work in. It used to be that we would scan as quickly as possible into a directory, named `to-import`, and import as quickly as possible into another named `imported`. This worked by using chokidar to watch files in `to-import` to pull into the queue of files to import. A while back that all changed and we pull sheets one at a time and process them before the scanner pulls in the next sheet. We no longer did any file watching and so the `imported` and `to-import` directories no longer really meant much.

This commit replaces that structure with a new one: we create a directory per batch and scan directly into those directories. Backups still put the files in a flat list, but that doesn't really matter as the database just refers to the right path in either case.

For development, this moves the scanned files out of `tmp` and into a folder called `dev-workspace`. However, the workspace can be configured using the `MODULE_SCAN_WORKSPACE` environment variable. In production we should set that to where we want it to be outside the source directory (which ideally should be read-only).