build(all): standardize on v4.1.3 of TypeScript
module-scan was the only project to require changes. One change was required by a stricter type parameter mapping to parameters. Maybe related to the tuple changes in recent TS versions? The other was required but IMO should not be. I filed it as https://github.com/microsoft/TypeScript/issues/41957.

The other version updates were for the issues react-scripts v4.0.0 had with verifying TypeScript config.