refactor(module-scan): simplify interpreter types
The class was named `SummaryBallotInterpreter`, and this made sense when it only interpreted summary ballots. I added the ability to interpret HMPB sheets before I understood what a "summary ballot" was.

This commit just renames it to `Interpreter`, which is also the name of the interface that it previously implemented. The interface didn't really do any good, so I just removed it.