build(ballot-encoder): fix package definition
When publishing 5.2.0 I realized too late that it only included the single `main` file in `lib`. This is because I forgot to update the `files` entry to include `lib` instead of `src`. This should now be fixed.

Additionally, `ts-emit-clean` didn't seem to be doing its job. Unsure why, but now that the files are in another directory it's trivial to clear them.