fix(scan): use score with best template alignment
Rather than assuming that the lowest score will be the one with the best template alignment, actually calculate the template alignment by seeing how many pixels would be newly black in the template. That is, minimize the number of pixels that are white in the scan but black in the template, as these clearly indicate a misalignment of some sort. Then, once we've found the offset that minimizes this error, use the score at that offset.

In practice this seems to raise all scores, even for empty bubbles. But it appears to raise the scores of filled bubbles more, creating a clearer gap between the two.