feat(bmd): add election hash to the BMD QR code
BMD election hash is a fixed length (20 hex characters), so we can omit the prefixed length typically used with strings. I could not make this change yet with HMPB election hash because we already have ballots printed for this election. Also, instead of encoding precinct and ballot style ids as strings, we now encode indexes as we do with HMPB metadata. This enabled sharing a decent amount of the logic for encoding and decoding BMD and HMPB data.