build(typescript): enable noUncheckedIndexedAccess
This only makes the change in ballot-encoder as a test. The rationale for this option is that TypeScript has a concept of optional properties which, when you access them, forces you to verify (or assert) that they're not undefined before you use them. This option is basically that, but for indexed access e.g. `precincts[0]`. Enabling it will make TS assume that the result of such accesses could be undefined.

Why wasn't it like this since strict null checks were enabled? Mostly because it can be annoying. JS developers rely on this being treated as not undefined a lot, so even now it's under a separate flag and not bundled into `strict`. I thought it worth trying out in a small library before trying it anywhere bigger.

https://devblogs.microsoft.com/typescript/announcing-typescript-4-1-beta/#no-unchecked-indexed-access