fix(scan): improve detection of blank pages
We look for blank pages before checking for QR codes because that is an expensive operation (~100s of ms). So we first check for a blank page and bail if we see that it's blank. Unfortunately, our blank page detection was essentially tallying all the would-be foreground pixels in a binarized version of the image using an automatic threshold. This means that for light/low-contrast images we get a high threshold (~240) introduce a significant amount of foreground pixels that are better counted as background.

To fix this, I now fix the blank page detection threshold at 200, which is low enough that it won't introduce a huge number of erroneous foreground pixels on a truly "blank" page. I also bump the allowed ratio of foreground/background for a blank page to be 0.07%. In practice, BMD front pages are about 1.3% at minimum, so this should leave us with a fairly comfortable margin.

An improvement to this would be to propagate the "blank"-ness of a page through the interpreter pipeline to prevent doing the thresholding and QR code detection again, potentially.