fix(scan): short circuit interpretation of blank pages
This is an improvement on #340. When the QR code detection step runs, it first determines whether the page is blank. If it is, checking for a QR code is skipped. However, we then do the blank page check and possibly QR code detection _again_ during interpretation. This commit passes the full results from the initial QR code detection so that we do not duplicate that work.

In my testing this does not noticeably speed up the processing of BMD ballots. I believe this is because the front side still takes longer (in parallel) to run than the back (blank) side due to loading/rotating/saving images. However, it's still a good cleanup.