feat(scan): check the election hash when scanning
We want to enforce that the scanned image has an election hash that matches the one the server was configured with. This isn't too bad implementation-wise. However, it exposed a lot of problems with our tests. They mostly depend on fixtures that are not set up to work when checking election hashes. For example, we can't just use `electionSample` from `@votingworks/fixtures` because it uses `asElectionDefinition` which uses the `JSON.stringify`d version of the election config that was imported, and therefore `JSON.parse`d, by node. This messes up the hash.

We could instead load the file from the built `@votingworks/fixtures`, right? Nope. That's wrong too because TypeScript reformats the JSON when building 🤦🏻‍♂️ (2-space indent to 4-space indent). So we instead have to load the file directly from where it lives in the source if we want to get the right hash. "Why don't we just hardcode the hash and give the server equivalent election config?", I hear you asking. Because the server actually validates that the election definition's string data hashes to the same value, so we can't do that.

On top of that, a lot of the fixtures in module-scan are built using older versions of the ballot. For example, they may be QR codes with the `ballot.page` URL that we never used for real and which don't have an election hash at all.

How do we fix this? Glad you asked:
- Add `electionData` to `ElectionDefinition` in `@votingworks/types`.
- Consolidate all `ElectionDefinition` types (I think there are 3).
- Add `electionData` everywhere we have an `ElectionDefinition`.
- Replace most uses of `asElectionDefinition` (and an even earlier incarnation I found in module-scan: `fromElection` in electionDefinition.ts) with a new `loadElectionDefinition` that reads from a file directly or something.
- Live with the fact that TypeScript modifies the files on build? Either that or stop using `.json` files and do something that preserves the raw bytes.
- Stop importing election definitions directly. It's convenient, but it's gonna bite us.
- Replace all old `ballot.page` fixtures with new ones.
- Surely more than this; it's not a trivial problem.