fix(scan): make election.json config work
When introducing the election hash check I inadvertantly broke scanning with just an election.json instead of a full ballot package. This is because that code path was `JSON.parse`ing the election.json before sending it to the server, thus changing the hash and preventing ballots from being scanned.

To address this, I changed a few things about how the scanner is configured:
- The API now expects `PATCH /config/electionDefinition` rather than `PATCH /config` with a nested object. This ensures we can get the raw JSON and hash it correctly.
- To remove an election config, the API expects `DELETE /config/electionDefinition`.
- Everything in the BSD election config path now expects an `ElectionDefinition`.
- Configuring by admin card is now removed from the scanner. It was just another code path to fix and we weren't using it anyway.

Additionally, there was a `fetch-mock` configuration issue that the changes to the tests brought to light. The simplest way to deal with it was just to use `jest-fetch-mock`.