perf(scan): improve scanner speed using JPEG
It seems the Fujitsu scanners scan noticeably faster on Linux with JPEG as the file format. I do not really understand why, but the results are pretty clear. Scanning sheets at 200dpi/gray/JPEG is a lot faster than 200dpi/gray/PNG. The difference is less notable with 300dpi/lineart.

Closes #364