feat(types): add `safeParse`
This upgrades the zod library to get a new feature: `ZodType#safeParse`. Instead of throwing an exception, failed parsing results in typed errors. To make this easier to use, I added `safeParseElection` and `safeParse` functions to `@votingworks/types`. These return a `Result`, which is a either `Ok` or `Err`, each containing a value representing success or failure. `Result` objects have some methods to make them easier to work with, a subset of the API for `Result` from the Rust language: https://doc.rust-lang.org/1.50.0/std/result/enum.Result.html.

I've been wanting something like this for a while to use instead of throwing exceptions, and this is my initial stab at it (if you don't count my previous half-hearted attempt in module-scan which this commit replaces).