fix(scan): improve the "delete batch" experience
Previously I had put in some code to quickly hide the batches that had been deleted while the server did the actual delete. However, this turned out to be flaky and would sometimes hide then show then hide the row. In practice deletes are really fast so having this optimization was not useful.

Additionally, the confirmation was using the built-in `confirm` function which doesn't provide a consistent UX with the rest of the application. This commit changes to use a `Modal` and to wait until the delete is confirmed to dismiss the modal.

Closes #366