refactor(types): improve schema refinement
zod is going through some restructuring, and v3 (in alpha) includes a feature that I really wanted before: custom dynamic error messages for refinements. It uses a new method on schemas called `superRefine` which may change, but for now is just what we need. It allows me to put the validation for various things, like enforcing unique ids, at the right place in the schema instead of at the root `Election`. This makes `safeParseElection` a pretty thin wrapper around the `Election` schema, whereas it wasn't before.