fix(bmd): simplify the timezone update behavior
Previously we used a JS Date and an IANA timezone string as the state for the admin screen modal for changing the date & time. Changing the timezone would change the date by whatever the difference in offsets was. This simplifies the behavior so that changing the timezone does not change the date, hopefully making it less likely to introduce accidental errors.

This commit uses the same library kiosk-browser uses for date & time representation, luxon.