fix(scan): update for latest backup format
`unzip` fails if one of the extract paths/globs doesn't match anything, so since we don't know off the bat whether it has JPEG or PNG files, just unzip everything.

Similarly, `sendFile` wants an absolute path and the paths in the database for image files are relative. This updates the returned paths to be absolute by assuming they're relative to the database file's directory.