refactor(scan): replace `scanSheets` return API
Rather than using an async generator function, just return an object with a domain-specific API. We weren't able to use the async generator function with a `for await` anymore anyway, so it didn't really justify its costs. Additionally, we're going to need to add to this API for controlling the precinct scanner.