fix(em): use non-broken `graceful-fs` version
https://github.com/isaacs/node-graceful-fs/commit/c55c1b8cb32510f92bd33d7c833364ecd3964dea changed `read.__proto__ = fs$read` to `Object.setPrototypeOf(read, fs$read)` to use the more standards-compliant `Object.setPrototypeOf`. However, that function fails if its second argument is `undefined`, which it is when running in the browser by default in a React app (`fs$read` is from `require('fs').read`, and the `fs` module is stubbed). This change was made in `graceful-fs@4.2.5`, and is present in `4.2.6` as well.

To work around this, we just pin to `4.2.4` for now. I'm not sure what the long-term solution should be. Perhaps `graceful-fs` should be graceful (hah!) about this scenario? Or maybe we could have webpack provide a more fully-featured stub to prevent this issue? Or maybe `archiver-utils` shouldn't be including `graceful-fs` at all in the browser?