fix(scan): cancel async operations on unmount
Addresses "Can't perform a React state update on an unmounted component.". In practice the fixes here are not anything that'd happen in production because `AppRoot` will never be unmounted. However, this does happen in tests and likely happens elsewhere in this and other apps. I'm adding this as a demo for how to deal with this problem.

https://dev.to/rodw1995/cancel-your-promises-when-a-component-unmounts-gkl