ci(all): improve pnpm caching
pnpm uses a central store to keep a single copy of each downloaded package version. When installing, it will reuse a specific package if present in the store, but if not it'll download it. Since we were not caching the store in CI, pnpm was always re-fetching the packages.

This changes the CircleCI config to cache just the store, so pnpm will always do the install but with everything already downloaded.