feat(types): add `AsyncResult`
The `Result` type was built to represent an operation that could succeed or fail, providing methods for working with that result and chaining it etc. This is good, except that many of the operations we may want to do are asynchronous. This meant we could not use the chaining methods such as `andThen` or `map` very effectively.

This commit adds `AsyncResult`, which has the same methods as `Result` except that they are promise-aware. This allows chaining operations that are asynchronous, with the terminating methods of `.ok` and `.err` etc. return promises instead of bare values.