style(all): configure format-on-commit
This uses `husky` v6 at the root of the workspace and `lint-staged` in each of the projects that need formatting. The lastest `husky` works a little differently than what we were using before (v4). Mostly, it uses the `core.hookspath` git config to point to `.husky/` within the project, setting that config property using a custom `prepare` script at the root. We then have a `pre-commit` script that asks `pnpm` to figure out all the changed packages and runs the `pre-commit` script in each one if it's there.