fix(lint): do not flag `Result` assignments
When assigning a `Result` to a variable as a statement, the type of that statement's expression is also `Result` because the value of an assignment is the value of the right-hand side. This triggered the `no-floating-results` rule incorrectly. Now it ignores assignments, assuming that the assigned `Result` will be used later.