feat(bmd): allow selection of All Precincts
Changes most `precinctId` values to be a `PrecinctSelection`, which can either be a single precinct or all precincts. Selecting "All Precincts" requires the cardless voting flow to select both a precinct and a ballot style. Because of that, I changed some of the terminology from e.g. "Activate Ballot Style" to "Activate Voter Session".

Additionally, I left the reports as close to "as-is" as possible. This may not be what we actually want for the "All Precincts" case, as it will print a single combined report. If we want it to print multiple reports we'll have to change it to do so.

Closes #498