refactor(em): move tally file reset out of modal
Though convenient from a code reuse perspective, putting this code in the modal meant that it was not really possible to perform an asynchronous tally update while also updating the state to avoid rendering the modal with now-invalid tally data. For example, removing the SEMS file and awaiting it properly would trigger a re-render of the modal where the `fileType` was still `SEMS`, but there was no longer and matching file to remove. This was cause an error when rendering.

This commit adds a new context function to do the reset properly, and that function is used alongside the modal to properly set the state to prevent the modal rendering as well as awaiting the tally update.