refactor(tally): use discriminated union
Rather than making a "base type" `CardTally` that contains the shared properties for the two "sub type" tally interfaces, just make `CardTally` a discriminated union. This enables us to reduce the amount of casting we need to do since TS can infer from the value of `tallyMachineType` which one we're working with.