test(ps): add previews
This adds support for viewing previews of various screens to make development easier. Sometimes it's hard to get to a certain state in the app and we have to temporarily hardcode things to make it work the way we want. Instead, this allows a developer to go to `/preview` when running the dev server and get a list of screens that can be previewed in various states. Those previews will live-update the same way anything else does in development.

This is related to #600, but isn't intended to be generic (yet) or show e.g. a component gallery. I decided to build this from scratch because it wasn't that much work and we can customize it to our needs. It follows a pattern similar to SwiftUI previews: https://www.avanderlee.com/swiftui/previews-different-states/.