feat(ps): give more fine-grained scanner status
The backend now tracks the current operation and returns a status appropriate to that operation, such as `Scanning`. In particular, when using the precinct scanner, the state of a sheet requiring review having been ejected toward the voter will now be represented as `Rejected` rather than `ReadyToScan`. This will make it much easier in a future change to have the frontend do less of its own state management.