fix(types): stop using `as` to the extent possible
My mental model for `as` is that it tells TypeScript "this value isn't quite like this type, but it's close enough so just let me use it kthxbai". In most of the cases we were using it that's not what we meant. Instead, we meant "I'd like to tell you what type this is so you can help me make sure it's right". But because `as` allows for a certain amount of fuzziness in the type matching, we weren't getting the same benefits as an actual type annotation.

So I went through and replaced `as` in most places with an appropriate replacement. Sometimes they could simply be removed with no reduction in type assistance. In places where we do need something, I am now using a trivial utility function called `typedAs` that simply acts as an inline type annotation on a value.