fix(all): fix python apps in `prod-build`
When running `make build` in the python apps by default `pipenv` installs dependencies in a shared folder. Doing that and then copying the contents of the project elsewhere means that the shared dependencies are no longer linked to the copy, so e.g. `flask` can't be found.

This fixes the issue by telling `pipenv` to put dependencies in the project so they can be copied alongside our code, similar to the default behavior of `node_modules`.