feat(scan): add write-in adjudication
This mostly prepares write-in adjudication for demonstration purposes. For now, it favors write-in adjudication over other possible reasons. Thus if a ballot page has both overvotes and write-ins, we'll get the new write-in adjudication screen. In the future we need something smarter that combines them. The write-ins provided by the user are, for now, ignored. We simply mark the page as adjudicated and move on.