feat(scan): record and use write-in adjudications
`bsd` now submits write-in adjudications to `module-scan`, which records them and uses them when exporting CVRs. Write-in CVR entries will now look like `__write-in-INDEX-NAME`, e.g. `__write-in-0-Bob Barker`. `election-manager` will lump anything that starts with `__write-in` into a single write-in tally, but we can change that later. Here's an example CVR with several adjudicated write-ins:

```json
{
  "775021789": ["__write-in-0-Lizard People"],
  "775021791": [
    "775033361",
    "775033362",
    "__write-in-0-Mario",
    "__write-in-2-Luigi"
  ],
  "775021795": ["775033091", "775033186", "__write-in-0-Bill & Ted"],
  "_ballotId": "f03d4622-4aaa-4265-90fa-e6bd024d925e",
  "_ballotStyleId": "2D",
  "_ballotType": "absentee",
  "_batchId": "424ff08b-195d-46a2-a4c8-f3408f4c7024",
  "_batchLabel": "Batch 10",
  "_precinctId": "775001535",
  "_scannerId": "000",
  "_testBallot": true,
  "_locales": { "primary": "en-US" },
  "_pageNumbers": [1, 2]
}
```