fix(all): correct dependencies of various hooks
Now that #928 fixed `useSmartcard` to return the same object if nothing has changed, we can fix the hooks that depend on its results to use the correct value (i.e. just `smartcard`). Doing so revealed a few other hooks with incorrect dependencies. This commit fixes those as well.

Some of them were such that, if you simply used the ESLint-suggested hook dependencies, you'd get stuck in an infinite-re-render loop. For example, in both BSD and EM there's a `useEffect` to update the current session. It used both `currentUserSession` and `setCurrentUserSession`. In some circumstances it would set the value and re-run the effect, updating the user session and running the effect again. The way out of this is to pass a function to `setCurrentUserSession` and to return the previous value in the situations where it wouldn't previously have been called.

Additionally, there was an unhandled `GET /machine-config` in EM's tests, so now that's handled by passing a machine config provider in the tests and always requiring it be passed to `AppRoot`.