fix(scan): retry scans for up to 2s
This was implemented as a change to `plustek-sdk`. The `#scan` client method now takes a predicate called `shouldRetry` to determine whether a retry should happen. Then, `module-scan` uses this to retry up to 2s.

This should fix an issue where the scanner sees a paper but can't actually grab it. It reports it is ready to scan, but actually scanning results in an IO error. This gives the scanner up to 2s to grab it before giving up.