fix(scan): reject mismatched pages
I attempted to do this with #589, but that merely triggered the need for adjudication of the sheet. Depending on what other reasons something might have for being adjudicated it could sneak through. Now when the sheet is about to be imported we validate it. If it doesn't pass, the interpretation will be replaced with `UnreadablePage` for both sheets, rendering it impossible to export.

Fixes #941