fix(bmd): do not enable screen reader on leaving admin screen
First, there's no need to enable/disable the screen reader. We can simply mute/unmute instead. Second, we don't want to unconditionally unmute. Instead, we want to restore the muted value as it was when we entered the admin screen. To do that, store the value in a `useEffect` closure.