fix(lint): make TS stop complaining
I've been getting errors, but only in VS Code, that the implicit type of the rules referenced a module that wouldn't be available to consumers of the library and that they are therefore not portable. I'm not sure I completely understand the error, but the fix seems to be to provide explicit type annotations for a few things.