refactor(scan): separate adjudication reason infos
Almost every time we want to look at the adjudication infos we actually just want the subset of them that correspond to the enabled adjudication reasons. This commit makes it so we now store separate lists for the enabled ones and the ignored ones, making it much harder to accidentally use adjudications reason infos that should be ignored.