test(ballot-encoder): use `fast-check`
This replaces my homegrown random fuzz tests with ones using `fast-check`. It's much more robust and actually did discover a bug: `writeUint` was doing bounds a check by bit-shifting, but this produced the wrong result when shifting by 31 (`1 << 31` is a negative number in JS because we're working with signed values).