fix(em): preserve scanner ID in CVR filename
When loading a CVR file EM shows the summary by extracting information from the filename. Because the filename has certain delimiters we have to be careful not to allow those characters to sneak in, so we sanitize the parts of the filename. However, this causes the scanner ID to be shown incorrectly because it may contain dashes (removed) and uppercase characters (converted to lowercase).

This commit fixes that by allowing correctly-formatted scanner IDs to be passed through as-is, sanitizing anything else. To make very sure the parsed values match the original values, this commit also adds a property test that tries many different values and ensures they all round-trip successfully.

Fixes #990