test(utils): add tests for round-trip filenames
This commit adds some property-based tests for the generate/parse functions for ballot package filenames. Doing so discovered a bug: if the election name was a value that would get sanitized to an empty string then the filename would have an extra double-underscore because the underscores on either side of the election name would be collapsed. This would cause the parser to fail because it found an unexpected number of sections separated by double-underscores.

I added a specific example covering that test failure and fixed it by using placeholder values when the value would otherwise be sanitized away. The property test is still present, but with constraints that mean it won't generate names that would be sanitized away.