feat(test-utils): add election arbitraries for property testing
## Context

An "arbitrary" is what `fast-check` calls a random value generator for a specific type of value. It ships with a bunch of basic ones like `fc.boolean()`, `fc.integer()`, `fc.string()`, etc. Arbitraries can be composed with some useful utilities like `fc.array()`, `fc.oneof()`, `fc.map()`, and `fc.chain()`.

When writing property-based tests you pass one or more `fc.Arbitrary` instances to `fc.assert` and it will call your check function with many generated values until it finds one that fails or it hits the maximum number of tries. If it finds a failing case it'll try to shrink it down to a minimal test case and show the error along with the minimal example.

## How this applies to vxsuite

We've got many different possible values for everything defining and election and the results and we can't easily test a wide variety of them. Property testing is one tool, in addition to the concrete hardcoded tests we have now, that allows us to widen the example space to find areas our code does not handle. It may mean that we write fewer tests with hardcoded values, but those will still have their place.

## What this commit adds

You'll now be able to import arbitraries for building entire `ElectionDefinition` objects or some part of them. This could be used  to test component rendering of election contests or metadata, to check tally code, etc.