fix(bmd): allow setting 1am or 1pm
This was broken by #828, a refactor that didn't properly update the hour value when moving from `0..11` to `1..12`. This fix includes a property test that checks lots of valid `DateTime`s to ensure they can all be set. I verified that without the fix the test found both 1am and 1pm as broken dates.

Fixes #1018