refactor(types): make `ContestId` a distinct type
## Context
This makes it impossible to use a `string` where a `ContestId` is expected. We do this by making `ContestId` be a string that _also_ has an extra unique property shared by no other types. This doesn't exist at runtime (where contest IDs are just strings), but we can use it to enforce parsing `string` into `ContestId`.

## Why?
Allowing the use of arbitrary strings for contest IDs can cause problems when the string is not a valid contest ID. We'd like it to fail as soon as it comes into the system and fails to parse as a contest ID, rather than causing potentially subtle issues later on.

A next step might be to add `CandidateId` and `WriteInId` types that better model what's allowed there, and would prevent us from using them interchangeably when they are not the same thing.

See https://timhwang21.gitbook.io/index/programming/typescript/newtype for some more why and how.