test(ui): fix `useAutocomplete` issues
The tests for `useAutocomplete` relied on a version of jest or react or some combination that made it prone to failing when unrelated package changes caused the version to change slightly. Rather than continuing to fight that, I'm upgrading `jest` and `ts-jest` myself now and dealing with the fallout. Specifically, I'm using legacy timers where required. I don't yet understand why they're needed, but they are. Also, I'm bringing `isWriting` in line with `isReading`, making it a ref rather than a state value.