refactor(bmd): pass reload down
Calling `window.location.reload` from `PollWorkerScreen` results in an error in our tests that is swallowed, but still printed to stderr:

> Error: Not implemented: navigation (except hash changes)

This is a limitation of JSDOM, but it's a sensible one. To address this problem, this commit modifies `App` to accept a `reload` prop that we mock in all the tests.