test(bmd): fix race condition in `TestBallotDeckScreen` test
When the test clicks "Print 63 ballots", it calls `handlePrinting` in `TestBallotDeckScreen`. That gets the printer info and determines whether a printer is connected. But the test was not waiting for the async stuff to complete, so sometimes it would before the `waitFor` looking for the error expired and sometimes it wouldn't.