refactor(router): replace `withRouter` with hooks
`react-router-dom` has `withRouter` that passes information via a higher-order component. This is fine, but it has fairly complex types that TS is only happy with when it's a default export. When we change to using named exports this will cause a build error. See https://github.com/microsoft/TypeScript/issues/42873.