feat(eslint): `gts-no-public-class-fields`
Requires that class properties be protected or private. This is part of the answer to satisfying this GTS rule:

> Mark properties that are never reassigned outside of the constructor with the `readonly` modifier (these need not be deeply immutable).

I decided to go with this approach because looking over the whole project from within  an ESLint rule to determine whether a property is ever updated or not is possible, but not trivial. Instead, I plan to make all properties private. This lets me enable the much easier-to-implement `@typescript-eslint/prefer-readonly` to achieve the aim of the GTS rule.