feat(types): add `unsafeParse` and `maybeParse` helpers
It's a little weird to have `safeParse(…).unsafeUnwrap()` when we don't ever intend to handle the failure "safely". Instead, it's better to just be explicit about it by calling `unsafeParse`, a helper function that does the same thing.

Similarly, `maybeParse` better communicates that we expect that the parse could fail, but that's fine and we just want the value if it parses successfully.